LPS-50919 move static util class and interface as well
