+++ portal-impl/src/com/liferay/portal/template/TemplateContextHelper.java
@@ -87,7 +87,7 @@ import com.liferay.portlet.expando.service.ExpandoColumnLocalService;
 import com.liferay.portlet.expando.service.ExpandoRowLocalService;
 import com.liferay.portlet.expando.service.ExpandoTableLocalService;
 import com.liferay.portlet.expando.service.ExpandoValueLocalService;
-import com.liferay.portlet.journalcontent.util.JournalContentUtil;
+import com.liferay.portlet.journal.util.JournalContentUtil;
 import com.liferay.taglib.util.VelocityTaglibImpl;
 
 import java.lang.reflect.Method;
