+++ portal-impl/src/com/liferay/portlet/journal/action/ExportArticleAction.java
@@ -36,7 +36,7 @@ import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.documentlibrary.util.DLUtil;
 import com.liferay.portlet.documentlibrary.util.DocumentConversionUtil;
 import com.liferay.portlet.journal.model.JournalArticleDisplay;
-import com.liferay.portlet.journalcontent.util.JournalContentUtil;
+import com.liferay.portlet.journal.util.JournalContentUtil;
 
 import java.io.File;
 import java.io.FileInputStream;
