+++ portal-impl/src/com/liferay/portlet/journalcontent/action/ViewAction.java
@@ -26,7 +26,7 @@ import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.journal.model.JournalArticle;
 import com.liferay.portlet.journal.model.JournalArticleDisplay;
 import com.liferay.portlet.journal.service.JournalArticleLocalServiceUtil;
-import com.liferay.portlet.journalcontent.util.JournalContentUtil;
+import com.liferay.portlet.journal.util.JournalContentUtil;
 
 import javax.portlet.PortletConfig;
 import javax.portlet.PortletPreferences;
